Pulling EOD from Yahoo using YahooFinance-Typhoeus gem. Storing it in CSV format for each Stock with TimeStamp. Changed the migration files so that the values are float.

+++ db/migrate/20120323082435_create_trading_days.rb
@@ -1,10 +1,10 @@
 class CreateTradingDays < ActiveRecord::Migration
   def change
     create_table :trading_days do |t|
-      t.integer :opening
-      t.integer :closing
-      t.integer :high
-      t.integer :low
+      t.float :opening
+      t.float :closing
+      t.float :high
+      t.float :low
       t.integer :volume
       t.date :date_of_trade
       t.references :stock
